easter egg added, now with proper dom referencing;
