+++ src/App.module.css
@@ -26,7 +26,3 @@
   display: none;
 }
 
-.easterEgg {
-  padding-top: 20px;
-  padding-left: 130px;
-}
