+++ src/components/KonamiCode/style.module.css
@@ -1,0 +1,32 @@
+.easterEgg {
+    display: none;
+    padding-top: 20px;
+    padding-left: 130px;
+  }
+
+.tortureDance {
+    padding-top: 20px;
+    padding-left: 130px;
+}
+
+.tortureDance {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    padding-top: 100px;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%; 
+    overflow: auto;
+    background-color: rgb(0,0,0);
+    background-color: rgba(0,0,0,0.4);
+}
+
+.modalContent {
+  background-color: #fefefe;
+  margin: auto;
+  padding: 20px;
+  border: 1px solid #888;
+  width: 80%;
+}
